+++ main.go
@@ -60,7 +60,7 @@ import (
 )
 
 // ReleaseVersion is the release version for the code.
-var ReleaseVersion = "0.7.4"
+var ReleaseVersion = "0.7.5"
 
 func main() {
 	os.Exit(main2())