+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 dev:
   - Add provider for epoch summaries
+  - Add withdrawals to epoch summaries table
 
 0.7.4:
   - Ensure that only canonical withdrawals are included in summaries